refactor: put in order component numbering and navigation, added Messages
