+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Three from "./components/Three";
-import { One } from "./components/One";
+import One from "./components/One";
 import { Two } from "./components/Two";
+import { Three } from "./components/Three";
 import { PrivateRoutes } from "./PrivateRoutes";
 import "./App.css";
 
@@ -12,9 +12,9 @@ const App = () => {
     <div className="App">
       <Router>
         <Routes>
-          <Route path={PrivateRoutes.PARAM_ONE} element={<Three />} />
-          <Route path={PrivateRoutes.PARAM_TWO} element={<One />} />
-          <Route path={PrivateRoutes.PARAM_THREE} element={<Two />} />
+          <Route path={PrivateRoutes.PARAM_ONE} element={<One />} />
+          <Route path={PrivateRoutes.PARAM_TWO} element={<Two />} />
+          <Route path={PrivateRoutes.PARAM_THREE} element={<Three />} />
         </Routes>
       </Router>
     </div>
