+++ src/Messages.ts
@@ -1,0 +1,15 @@
+export const MessageOne = {
+  START: "You are good to go!",
+  TITLE: "Happy Hacking",
+  BUTTON: "The Way",
+};
+
+export const MessageTwo = {
+  HINT: "Not everything is as it seems... you should inspect this situation.",
+  PEP: "You are almost there",
+};
+
+export const MessageThree = {
+  GIFT: "Download me and find the gift inside.",
+  GITHUB: "https://github.com/alestojanovic/scavenger-hunt",
+};
